python/turn_off_cluster/requirements.txt
azure-identity
azure-mgmt-containerservice